添加评论回复发送模板消息，添加根据id获取帖子详情接口

+++ src/main/java/com/lbxy/dao/ImageDao.java
@@ -18,7 +18,7 @@ public class ImageDao {
         return image.save();
     }
 
-    public List<Image> getImagesByContentIdAndType(int contentId,int type) {
+    public List<Image> getImagesByContentIdAndType(long contentId,int type) {
         return Image.DAO.find("select * from image where contentId = ? and type = ?", contentId, type);
     }
